+++ app/core/constants.py
@@ -3,12 +3,12 @@ SIMILARITY_THRESHOLD = 0.5
 KNOWLEDGE_BASE_URLS = [
     "https://thewanderbug.com/best-restaurants-melbourne/",
     "https://www.theworlds50best.com/discovery/sitemap/australia/melbourne",
-    "https://cluboenologique.com/story/melbourne-restaurants/",
-    "https://www.broadsheet.com.au/melbourne/food-and-drink/article/best-restaurants-melbourne",
-    "https://www.tripadvisor.com.au/Restaurants-g255100-zfp43-Melbourne_Victoria.html",
+    "https://whatson.melbourne.vic.gov.au/article/the-best-accessible-restaurants-in-melbourne/",
+    "https://www.gourmettraveller.com.au/dining-out/restaurant-guide/best-restaurants-melbourne-4478/",
     "https://www.theurbanlist.com/melbourne/a-list/best-restaurants-melbourne",
     "https://www.theage.com.au/goodfood/melbourne-eating-out/25-of-melbourne-s-most-iconic-restaurants-to-tick-off-your-bucket-list-20231107-p5ei86.html",
     "https://www.australiantraveller.com/vic/melbourne/melbourne-restaurants/",
     "https://www.eater.com/maps/best-melbourne-restaurants-38",
     "https://www.theinfatuation.com/melbourne/guides/best-melbourne-restaurants",
+    "https://www.gourmettraveller.com.au/dining-out/food-news/new-restaurants-melbourne-21063/",
 ]